Make Pod a pure marker trait
**This is a hard breaking change!**

Making Pod a marker trait and providing its methods as crate functions enables the rust semver trick.
This trick is impossible to execute in the current state and prevents any kind of change in design of this crate.
Bikeshedding the helper method names is now possible without further breaking changes!

At the same time I feel confident enough to embark on a 1.0 journey with the ability to improve the API if needed.